requirements.txt
loguru==0.7.3
RPi.GPIO==0.7.1
spidev==3.7
